requirements.txt
os  # This is a standard library; no need to include in requirements.txt
pinecone-client==2.2.2  # Replace with the latest version of Pinecone client
sentence-transformers==2.2.2  # For SentenceTransformer embeddings
langchain==0.0.303  # Ensure it supports ChatGroq and RetrievalQA
langchain-groq==0.1.2  # Hypothetical version for LangChain-Groq; replace as needed
langchain-community==0.0.3  # Replace with the appropriate version for community modules
langchain-huggingface==0.0.5  # Replace with the appropriate version for HuggingFace integration
torch==2.0.1  # Required by sentence-transformers
transformers==4.27.4  # Needed for HuggingFace models
numpy==1.24.3  # For numerical operations in embeddings